+++ cedar-policy/CHANGELOG.md
@@ -14,6 +14,11 @@ Starting with version 3.2.4, changes marked with a star (*) are _language breaki
 
 Cedar Language Version: TBD
 
+### Changed
+
+- For the `tpe` experimental feature, `PartialEntities::from_concrete` now requires a `Schema` and will validate the entities,
+  ensuring that a `PartialEntities` object always meets the preconditions required for type aware partial evaluation.
+
 ### Fixed
 
 - Removed incorrect dependency of feature `partial-eval` of feature `tpe`. (#1898)